+++ pkg/scheduler/plugins/proportion/reclaimable/strategies/strategies.go
@@ -7,7 +7,6 @@ import (
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/resource_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/log"
 	rs "github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/plugins/proportion/resource_share"
-	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/plugins/proportion/utils"
 )
 
 type MaintainFairShareStrategy struct{}
@@ -49,11 +48,13 @@ func (mfss *MaintainFairShareStrategy) Reclaimable(
 	reclaimeeRemainingShare rs.ResourceQuantities) bool {
 	// This strategy allows to reclaim if reclaimee is currently over allowed fair share
 
-	log.InfraLogger.V(6).Infof("Checking if reclaim is possible for reclaimer <%s> and reclaimee <%s> in order "+
-		"to maintain fair share. Reclaimee requested: <%s>, deserved: <%s>, fairShare: <%s>, "+
-		"reclaimeeRemainingShare: <%s>",
-		reclaimerQueue.Name, reclaimeeQueue.Name, reclaimeeQueue.GetRequestableShare(), reclaimeeQueue.GetDeservedShare(),
-		reclaimeeQueue.GetFairShare(), reclaimeeRemainingShare)
+	log.InfraLogger.V(6).Do(func() {
+		log.InfraLogger.V(6).Infof("Checking if reclaim is possible for reclaimer <%s> and reclaimee <%s> in order "+
+			"to maintain fair share. Reclaimee requested: <%s>, deserved: <%s>, fairShare: <%s>, "+
+			"reclaimeeRemainingShare: <%s>",
+			reclaimerQueue.Name, reclaimeeQueue.Name, reclaimeeQueue.GetRequestableShare(), reclaimeeQueue.GetDeservedShare(),
+			reclaimeeQueue.GetFairShare(), reclaimeeRemainingShare)
+	})
 
 	return FitsMaintainFairShare(reclaimeeQueue, reclaimeeRemainingShare)
 }
@@ -66,13 +67,15 @@ func (gdqs *GuaranteeDeservedQuotaStrategy) Reclaimable(
 	reclaimeeRemainingShare rs.ResourceQuantities) bool {
 	// This strategy allows to reclaim if reclaimer is under deserved quota ("starved") and reclaimer is above quota
 
-	log.InfraLogger.V(6).Infof("Checking if reclaim is possible for reclaimer <%s> and reclaimee <%s> in order to "+
-		"Guarantee deserved quota. "+
-		"Reclaimee requested: <%s>, deserved: <%s>, fairShare: <%s>, reclaimeeRemainingShare: <%s> "+
-		"Reclaimer requested: <%s>, deserved: <%s>, fairShare: <%s>",
-		reclaimerQueue.Name, reclaimeeQueue.Name, reclaimeeQueue.GetRequestableShare(), reclaimeeQueue.GetDeservedShare(),
-		reclaimeeQueue.GetFairShare(), reclaimeeRemainingShare, reclaimerQueue.GetRequestableShare(),
-		reclaimerQueue.GetDeservedShare(), reclaimerQueue.GetFairShare())
+	log.InfraLogger.V(6).Do(func() {
+		log.InfraLogger.V(6).Infof("Checking if reclaim is possible for reclaimer <%s> and reclaimee <%s> in order to "+
+			"Guarantee deserved quota. "+
+			"Reclaimee requested: <%s>, deserved: <%s>, fairShare: <%s>, reclaimeeRemainingShare: <%s> "+
+			"Reclaimer requested: <%s>, deserved: <%s>, fairShare: <%s>",
+			reclaimerQueue.Name, reclaimeeQueue.Name, reclaimeeQueue.GetRequestableShare(), reclaimeeQueue.GetDeservedShare(),
+			reclaimeeQueue.GetFairShare(), reclaimeeRemainingShare, reclaimerQueue.GetRequestableShare(),
+			reclaimerQueue.GetDeservedShare(), reclaimerQueue.GetFairShare())
+	})
 
 	// reclaimer has to be under (or equal) deserved quota in all resources (cpu, mem, gpu)
 	if !ReclaimerFitsDeservedQuota(reclaimerResources, vectorMap, reclaimerQueue) {
@@ -89,7 +92,7 @@ func (gdqs *GuaranteeDeservedQuotaStrategy) Reclaimable(
 
 // FitsMaintainFairShare returns true when the reclaimee remains over its allocatable share.
 func FitsMaintainFairShare(reclaimeeQueue *rs.QueueAttributes, reclaimeeRemainingShare rs.ResourceQuantities) bool {
-	return !reclaimeeRemainingShare.LessEqual(reclaimeeQueue.GetAllocatableShare())
+	return !reclaimeeQueue.QuantitiesLessEqualAllocatable(reclaimeeRemainingShare)
 }
 
 // ReclaimerFitsDeservedQuota returns true when adding the reclaimer keeps its queue within deserved quota.
@@ -98,10 +101,7 @@ func ReclaimerFitsDeservedQuota(
 	vectorMap *resource_info.ResourceVectorMap,
 	reclaimerQueue *rs.QueueAttributes,
 ) bool {
-	reclaimerRequestedQuota := reclaimerQueue.GetAllocatedShare()
-	reclaimerRequestedQuota.Add(utils.QuantifyVector(reclaimerResources, vectorMap))
-
-	return reclaimerRequestedQuota.LessEqual(reclaimerQueue.GetDeservedShare())
+	return reclaimerQueue.AllocatedPlusResourcesLessEqualDeserved(reclaimerResources, vectorMap)
 }
 
 // ReclaimeeExceedsDeservedQuota returns true when the reclaimee remains over deserved quota.